Move implementation to separate module
